Add support for Font Awesome 5
Font Awesome 5 adds four new prefixes:
* `fab` for brands
* `fas` for the original solid style
* `far` for regular outline style
* `fal` for light outline style

The actual prefix on the icon name is still `fa-`. So, for example, the Github icon (a brand) is `”fab fa-github”`.

For backward compatibility, the `fa` prefix is still supported, and acts as an alias for `fas`, which means only the solid style icons are available.

Fixes #8.